Separate lz77, derive default UintConfig from HistogramParams, BuildAndStoreEntropyCodes and BuildAndStoreANSEncodingData into EntropyEncodingData members
