+++ lib/lib.gni
@@ -359,6 +359,8 @@ libjxl_enc_sources = [
     "jxl/enc_image_bundle.h",
     "jxl/enc_linalg.cc",
     "jxl/enc_linalg.h",
+    "jxl/enc_lz77.cc",
+    "jxl/enc_lz77.h",
     "jxl/enc_modular.cc",
     "jxl/enc_modular.h",
     "jxl/enc_noise.cc",
